+++ AmandsSense/Properties/AssemblyInfo.cs
@@ -4,7 +4,7 @@
 [assembly: AssemblyCompany("")]
 [assembly: AssemblyProduct("AmandsSense")]
 [assembly: AssemblyCopyright("Copyright © 2022-2025 Amanda. All rights reserved.")]
-[assembly: AssemblyTrademark("SPT 3.11.0")]
+[assembly: AssemblyTrademark("SPT 3.11.1, EFT 0.16.1.3.35392")]
 
-[assembly: AssemblyVersion("2.0.1.0")]
-[assembly: AssemblyFileVersion("2.0.1.0")]
+[assembly: AssemblyVersion("2.0.1.1")]
+[assembly: AssemblyFileVersion("2.0.1.1")]
